requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
numpy>=1.26.0
Pillow>=10.0.0
matplotlib>=3.8.0
tqdm>=4.66.0
jupyter>=1.0.0
